+++ app-admin/src/environments/environment.prod.ts
@@ -5,8 +5,5 @@ export const environment: iEnv = {
 	dev: false,
 	production: true,
 	debug: false,
-	api: "https://api.bolaopenacova.com",
-	analytics_api: "https://analytics.bolaopenacova.com",
 	title: "Bolão Pé na Cova 3.0",
-	web: "https://bolaopenacova.com",
 };