+++ app-admin/src/environments/environment.staging.ts
@@ -5,8 +5,5 @@ export const environment: iEnv = {
 	dev: true,
 	production: true,
 	debug: false,
-	api: "https://dev.api.bolaopenacova.com",
-	analytics_api: "https://dev.analytics.bolaopenacova.com",
 	title: "Bolão Pé na Cova 3.0",
-	web: "https://dev.bolaopenacova.com",
 };