+++ src/changelog.md
@@ -1,3 +1,9 @@
+## 2.1.1
+2026-06
+	- **new:** angular admin api dynamic
+	- **fix:** logo as default fix
+	- **improvement:** deployment process improvement
+
 ## 2.1.0
 2025-10
 	- **new:** changelog